[webapp] rename Telegram context files
