+++ services/webapp/ui/src/contexts/TelegramProvider.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { useTelegram } from "@/hooks/useTelegram";
-import { TelegramContext } from "./telegram-context";
+import { TelegramContext } from "./telegramContext";
 
 export const TelegramProvider = ({ children }: { children: ReactNode }) => {
   const telegram = useTelegram();
