+++ services/webapp/ui/src/pages/Analytics.tsx
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import { MedicalHeader } from '@/components/MedicalHeader';
 import { ChartContainer, ChartTooltip, ChartTooltipContent } from '@/components/ui/chart';
 import { fetchAnalytics, fallbackAnalytics } from '@/api/stats';
-import { useTelegramContext } from '@/contexts/telegram-context';
+import { useTelegramContext } from '@/contexts/telegramContext';
 
 const Analytics = () => {
   const navigate = useNavigate();
